lab2/code/requirements.txt
PyQt6==6.6.1
Pillow==10.0.1
python-dateutil==2.8.2
